Add ITunerHostManager service
